Img_Upscaler/requirements.txt
opencv-contrib-python-headless
flask
Pillow